use cookie for storing session token, and set by JS code
As I've learned, passing a token in an "Authenticate: Bearer ..." header
is possible only if some client-side Javascript creates that header;
the browser itself cannot do that.
Therefore switch over to a cookie for storing the token. It's created
by JavaScript code on successful authentication, and parsed by the
request handler.